+++ source/toy_bucket.h
@@ -11,12 +11,12 @@
 //a custom allocator
 typedef struct Toy_Bucket {  //32 | 64 BITNESS
 	struct Toy_Bucket* next; //4  | 8
-	size_t capacity;         //4  | 4
-	size_t count;            //4  | 4
+	unsigned int capacity;   //4  | 4
+	unsigned int count;      //4  | 4
 	char data[];             //-  | -
 } Toy_Bucket;                //12 | 16
 
-TOY_API Toy_Bucket* Toy_allocateBucket(size_t capacity);
-TOY_API void* Toy_partitionBucket(Toy_Bucket** bucketHandle, size_t amount);
+TOY_API Toy_Bucket* Toy_allocateBucket(unsigned int capacity);
+TOY_API void* Toy_partitionBucket(Toy_Bucket** bucketHandle, unsigned int amount);
 TOY_API void Toy_freeBucket(Toy_Bucket** bucketHandle);
